Changed IRC parser to confirm to IRC Protocol standards and accept any characters for usermode.
Added support for IRCOP / Opers for User Levels.

Changed "Settings" to "Configuration" Settings.xml is now defined as app.config
and the "Settings" node in the configuration file should be "Config".

Standardised the API for console output to a single function.

Made the posting of IRC server traffic to console optional to circumvent UNIX
system logs being loaded with garbage.